video-editor/pyproject.toml
[project]
name = "video-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.54.1",
    "python-dotenv>=1.2.2",
]
